third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/popinContextType.in-popin.tentative.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify PopinContextType is 'partitioned' in a Partitioned Popin
  assert_equals: expected "partitioned" but got "popinContextType is not supported"
Harness: the test ran to completion.
